resolve last hammer circ dep
