+++ src/RNGestureHandlerModule.web.ts
@@ -7,6 +7,7 @@ import type { Config } from './web/interfaces';
 import InteractionManager from './web/tools/InteractionManager';
 import NodeManager from './web/tools/NodeManager';
 import * as HammerNodeManager from './web_hammer/NodeManager';
+import * as createGestureHandler from './web_hammer/modifyGestureHandlerRegistry';
 import { GestureHandlerWebDelegate } from './web/tools/GestureHandlerWebDelegate';
 
 export default {
@@ -48,7 +49,7 @@ export default {
       // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
       const GestureClass = HammerGestures[handlerName];
       // eslint-disable-next-line @typescript-eslint/no-unsafe-call
-      HammerNodeManager.createGestureHandler(handlerTag, new GestureClass());
+      createGestureHandler.createGestureHandler(handlerTag, new GestureClass());
     }
 
     this.updateGestureHandler(handlerTag, config as unknown as Config);
@@ -97,7 +98,7 @@ export default {
     if (isNewWebImplementationEnabled()) {
       NodeManager.dropGestureHandler(handlerTag);
     } else {
-      HammerNodeManager.dropGestureHandler(handlerTag);
+      createGestureHandler.dropGestureHandler(handlerTag);
     }
   },
   // eslint-disable-next-line @typescript-eslint/no-empty-function
